Initial commit of smart creation of power cap functions
Uses estimates of interesting power levels (around the amount needed to
finish) and times (to be interesting any dynamics in the power cap need
to happen before the run ends, and the time periods involved must be
comparable to the makespan).

Times are based on the computation time for the critical path of the
DAG.

Power cap sizes are based on the energy used for an unconstrained run
using HEFT and this minimum finish time estimation.